AFT: Change "machine" words to "model" or "device"
Use of the word "machine" has been incohesive in the code. Sometimes it
means the model of the device e.g. gigabyte and sometimes it means the
spesific device name e.g. gigabyte_1. Model and device words are used
more in the code anyway so changing "machine" words to either "model"
or "device" makes the code more understandable.

Signed-off-by: Simo Kuusela <simo.kuusela@intel.com>