AFT: Beaglebone harness dnsmasq support
In AFT /usr/lib/misc/dnsmasq.leases file is used to find the DUT ip
address using the device id given in AFT device config files. In normal
setup the id is always the same with the same device. For some reason
when using Beaglebone Black as testing harness the device id in the
dnsmasq.leases file changes every time it gives the DUT an ip address.
This is why we have to take the first ip address in the dnsmasq.leases
file because we cant use the device id to identify the right one.
Anyways Beaglebone harness will always have only one DUT connected to it
so taking always the first ip address should work fine.

Signed-off-by: Simo Kuusela <simo.kuusela@intel.com>